requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography>=45.0.7
daphne==4.2.1
deepl==1.22.0
distro==1.9.0
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jiter==0.10.0
jsonfield==3.2.0
kombu==5.5.4
Markdown==3.8.2
msgpack==1.1.1
multidict==6.6.4
openai==1.101.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.4.0
tqdm==4.67.1
twilio==9.8.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
yarl==1.20.1
zope.interface==8.0.1
